сделана большая часть задания
